+++ src/Outro.js
@@ -1,0 +1,53 @@
+const Outro = () => {
+  return (
+    <div>
+                 <h3 style={{ lineHeight: "2.4rem", fontSize:"2.2rem" }}>
+             ~ Outro ~ 
+            </h3>
+<br></br>
+            Expressions and acknowledgements of peace, for a mutual assured
+            regeneration,{" "}
+            <strong>is all about standing equal in peace and trust making. </strong>
+     <br></br>     <br></br>
+
+      What we have been observing so far: the ones opts-in and out of peace
+      treaties, traditionally and naturally, by delegation or inheritance of power, have been
+      the ones with highest capacity of violence, whether they use it or not in an offensive manner. <br></br>
+      <br></br> However the majority impacted by conflicts, have no significant
+      labor, direct impact, hence legitimacy in the making and sustaining a world
+      peace.
+      <br></br>
+      <br></br>
+
+      
+      It is time to separate power and violence, in fact, search power somewhere
+      else: in the conscious and wisdom of a great civilisation:
+      <br></br>
+      <strong>#GenerationPeace</strong>
+      <br></br> <br></br>
+      <a
+        href="https://www.youtube.com/watch?v=xxLLIxuY3gg"
+        target={"_blank"}
+        rel="noreferrer"
+      >
+        <u>Here</u>
+      </a>{" "}
+      is a short video and{" "}
+      <a
+        href="https://www.e-ir.info/2012/10/09/the-social-contract-theory-in-a-global-context/"
+        target={"_blank"}
+        rel="noreferrer"
+      >
+        <u>here</u>
+      </a>{" "}
+      is an article about social contract theories and their evolution/adoption throughout the history.
+      <br></br>
+      <br></br>
+      <hr></hr>
+      <br></br>
+      <br></br>
+    </div>
+  );
+};
+
+export default Outro;
