+++ src/Reasoning.js
@@ -1,0 +1,39 @@
+const Reasoning = () => {
+    return (
+        <div>
+          <img
+              alt="Anakin and Padme in love with world peace by a petition"
+              src="anakin-padme.png"
+              width={"60%"}
+            ></img>
+                      {/* <a href="https://imgflip.com/i/6egiif"><img src="https://i.imgflip.com/6egiif.jpg" title="made at imgflip.com"/></a><div><a href="https://imgflip.com/memegenerator">from Imgflip Meme Generator</a></div> */}
+
+<br></br>
+<br></br>
+             <strong>
+              That's why we'd like this whole process to be treated, perceived
+              and implemented as a global petition,{" "}
+            </strong>{" "}
+            where people, in essence, proactively encouraging and empowering
+            their respective states to reflect and align with their wishes and
+            provide a sustainable security.
+            <br></br>
+            <br></br>
+            This time, with a shield of an ultimate peace, rather than war
+            machines,{" "}
+            <strong>
+              {" "}
+              <i> first time in the history</i>.
+            </strong>
+            <br></br> <br></br>
+
+            In the making of a World Peace, via experiencing of an emerging
+            human right, it would be a fallacy to make it a money game.
+            <br></br> <br></br>
+            We are designing incentives and a gamified experience however,
+            expressions and acknowledgements will never have a cost or
+            over-requested data.
+         
+        </div>
+    ); }
+export default Reasoning;
