+++ src/human-right.css
@@ -1,0 +1,4 @@
+.human-right {
+    background: black;
+    color: white;
+}
